3619: Rename EntityAttributeItemDelegate files
