+++ common/src/View/EntityPropertyGrid.cpp
@@ -22,9 +22,9 @@
 #include "Macros.h"
 #include "Model/EntityProperties.h"
 #include "View/BorderLine.h"
-#include "View/EntityAttributeItemDelegate.h"
 #include "View/EntityAttributeModel.h"
 #include "View/EntityAttributeTable.h"
+#include "View/EntityPropertyItemDelegate.h"
 #include "View/MapDocument.h"
 #include "View/ViewConstants.h"
 #include "View/QtUtils.h"
